requirements/inference.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
librosa==0.10.1
python-dotenv
backcall==0.2.0
backports.cached-property==1.0.2
beautifulsoup4==4.12.2
bleach==6.0.0
botocore==1.32.5
build==0.10.0
CacheControl==0.12.14
cachetools==5.3.2
cffi==1.15.1
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
crashtest==0.4.1
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
dulwich==0.21.6
entrypoints==0.4
exceptiongroup==1.2.0
fastjsonschema==2.19.0
filelock==3.12.2
flatbuffers==23.5.26
frozenlist==1.3.3
fsspec==2023.1.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
html5lib==1.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
jaraco.classes==3.2.3
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
jupyter-console==6.6.3
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
keyring==23.13.1
kubeflow==0.0.1rc0
lightning-utilities==0.10.0
lockfile==0.12.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.5.8
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.21.6
oauthlib==3.2.2
onnxruntime==1.14.1
packaging==23.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip-tools==6.12.2
pipfreeze==2.0.3
pkginfo==1.9.6
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
pluggy==1.2.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
prometheus-client==0.17.1
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
pyngrok==7.0.1
pyproject_hooks==1.0.0
pyreadline==2.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytorch-lightning==1.9.5
pywin32-ctypes==0.2.2
pyzmq==24.0.1
qtconsole==5.4.4
QtPy==2.4.1
rapidfuzz==3.4.0
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rsa==4.9
torchvision
diffusers
sentry-sdk==1.36.0
setproctitle==1.3.3
pillow
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
slurm==0.4.6
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4.1
sympy==1.10.1
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.3
torch==1.13.1
torchmetrics==0.11.4
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
trove-classifiers==2023.11.14
typing_extensions==4.7.1
urllib3==1.26.18
virtualenv==20.24.7
wandb==0.16.0
webencodings==0.5.1
websocket-client==1.6.1
widgetsnbextension==4.0.9
wincertstore==0.2
yarl==1.9.3
zipp==3.15.0
